So we start filling out the function.
Remember that StringEngine takes and ARRAY of Strings. Clojure doesn't have quite as good support for arrays as it does for maps, sets, lists and vectors. I believe this is because arrays don't have type erasure.

Anyway, a simple way to create an array is with (into-array). However, we need to tell it what parameterized type we want, thus the 'String' as the second param.