Here's a really basic proxy call.
The first vector is a list of at most one class, which must be first, and N interfaces, which you want to extend.

The second vector (here, it's empty), is a set of arguments to the superclass constructor.

Finally, each method you want to override is specified as you see. 'this' is implicit, as it is in Java.