In cases like this, where you only have interfaces and not concrete or abstract classes to extend, reify is considered a better choice.
You CAN extend methods from Object as well as implement anything your interface calls for.

The difference between reify and proxy are interesting. With proxy, you had a list of classes/interfaces to extend and a list of constructor params.

With reify, you have a bunch of 'specs' - the name of an interface followed by method implementations.

Also, reify requires that you explicitly write the 'this' parameter in, though you shouldn't add it to any recur calls.