Looking at defrecord is easier if we look at defprotocol as well.
And because I'm an idiot, I've pulled in a math library as well.

A class created by defrecord has two factory methods: ->Type and map->Type. Here we're using ->TwoDCartesian in our tests, which requires that we pass the field values we've specified in defrecord.

Accessing the fields can be done in several ways. For 'this', you can simply access them directly, as you would in any Java class.

Externally, you can use the field access notation we saw a while ago: '(.field obj)'.