This is a little embarrassing.
I looked and I looked, but I couldn't find a way to load a CLASS created by defrecord through the clojure.java.api.Clojure API.

I BELIEVE it has something to do with when the class is generated.

I KNOW that the namespace is a bit munged for the generated class.